server/requirements-all.txt
# Consolidated Requirements for All Phases
# =========================================
# This file includes all dependencies for the complete system

# Flask and Core Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0

# Database
supabase==2.3.0
postgrest==0.13.2

# GitHub Integration
PyGithub==2.1.1
github3.py==4.0.1

# Docker
docker==7.0.0

# Phase 1: Logging
python-json-logger==2.0.7

# Phase 1: Queue System
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Phase 1: Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
requests-mock==1.11.0
faker==20.1.0

# Phase 3: Security
cryptography==41.0.7

# Phase 3: Monitoring
sentry-sdk[flask]==1.39.1

# Phase 2/4: Agent Selection
numpy==1.26.2

# Utilities
urllib3==2.1.0
